Move function: `Model::from_path` => `input::load_model`
